feat: split area.rs into a module
